+++ lib/src/_internals/route_menu.dart
@@ -18,7 +18,7 @@ class RoutePullDownMenu extends StatelessWidget {
     required this.routeTheme,
     required this.alignment,
     required this.animation,
-    required this.initialScrollOffset,
+    required this.scrollController,
   });
 
   /// Items to show in the menu.
@@ -37,7 +37,7 @@ class RoutePullDownMenu extends StatelessWidget {
   final Alignment alignment;
 
   /// Is used to define the initial scroll offset of menu's body.
-  final double? initialScrollOffset;
+  final ScrollController? scrollController;
 
   @override
   Widget build(BuildContext context) {
@@ -85,7 +85,7 @@ class RoutePullDownMenu extends StatelessWidget {
                   axisAlignment: -1,
                   sizeFactor: clampedAnimation,
                   child: MenuBody(
-                    initialScrollOffset: initialScrollOffset,
+                    scrollController: scrollController,
                     items: items,
                   ),
                 ),